+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/offer/Currency.java
@@ -1,0 +1,16 @@
+package pl.com.bottega.ecommerce.sales.domain.offer;
+
+public class Currency {
+
+    private String currency;
+    private String currencyCode;
+
+    public String getCurrency() {
+        return currency;
+    }
+
+    public String getCurrencyCode() {
+
+        return currencyCode;
+    }
+}
