+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/offer/Discount.java
@@ -1,0 +1,30 @@
+package pl.com.bottega.ecommerce.sales.domain.offer;
+
+import java.math.BigDecimal;
+
+public class Discount {
+
+    private String discountCause;
+
+    private BigDecimal discount;
+
+    public Discount(BigDecimal discount, String discountCause){
+        this.discount = discount;
+        this.discountCause = discountCause;
+    }
+
+    public BigDecimal getDiscount() {
+        return discount;
+    }
+
+    public String getDiscountCause() {
+        return discountCause;
+    }
+
+    @Override public int hashCode() {
+        final int prime = 31;
+        int result = 1;
+        result = prime * result + (discount == null ? 0 : discount.hashCode());
+        return result;
+    }
+}
